[feat]: remove tests, split clasess in different files and rework classes
