+++ src/LLM/__init__.py
@@ -1,2 +1,4 @@
-from Edulytica.src.LLM.IModel import *
-from Edulytica.src.LLM.utils import *
+from src.LLM.IModel import *
+from src.LLM.utils import *
+from src.LLM.Model_instruct import *
+from src.LLM.Model_pipeline import *
