+++ src/LLM/vikhr/Vikhr_Nemo_instruct.py
@@ -1,67 +1,17 @@
-from transformers import AutoModelForCausalLM, AutoTokenizer, GenerationConfig, pipeline
-from Edulytica.src.LLM import IModel
-from Edulytica.src.LLM import DEFAULT_SYSTEM_PROMPT
+from transformers import AutoModelForCausalLM, AutoTokenizer, GenerationConfig
+from src.LLM import Model_instruct
+from src.LLM import DEFAULT_SYSTEM_PROMPT
 
 
-class Vikhr_Nemo_instruct(IModel):
-    """Class for Vikhr-Nemo instruct usage"""
+class Vikhr_Nemo_instruct(Model_instruct):
+    """Class for Vikhr-Nemo instruct usage. For inference using huggingface transformers"""
 
-    def __init__(self, model_name=None, system_prompt=DEFAULT_SYSTEM_PROMPT, device_map="auto"):
+    def __init__(self, model_name=None, chat_template=None, system_prompt=DEFAULT_SYSTEM_PROMPT, device_map="auto"):
         if model_name is None:
             model_name = "Vikhrmodels/Vikhr-Nemo-12B-Instruct-R-21-09-24"
-        self.tokenizer = AutoTokenizer.from_pretrained(model_name)
-        self.model = AutoModelForCausalLM.from_pretrained(model_name, device_map=device_map)
-        self.device = self.model.device
-        self.system_prompt = system_prompt
-
-    def apply_chat_template(self, system_prompt: str, prompts: list):
-        """Method for applying chat template"""
-        messages = list()
-        for prompt in prompts:
-            message = [
+        if chat_template is None:
+            chat_template = lambda prompt: [
                 {"role": "system", "content": system_prompt},
                 {"role": "user", "content": prompt}
             ]
-            messages.append(message)
-        texts = list(
-            map(lambda message: self.tokenizer.apply_chat_template(message, tokenize=False, add_generation_prompt=True),
-                messages))
-        return texts
-
-    def generate(self, prompts: list, max_new_tokens=512):
-        """Method for text generation by model.
-        Takes list of prompts and max new tokens and return list with model generated text"""
-        generation_config = GenerationConfig(max_new_tokens=max_new_tokens)
-
-        texts = self.apply_chat_template(self.system_prompt, prompts)
-        model_inputs = self.tokenizer(texts, return_tensors="pt").to(self.device)
-
-        outputs = self.model.generate(**model_inputs, generation_config=generation_config)
-        response = self.tokenizer.batch_decode(outputs, skip_special_tokens=True)
-        return response
-
-
-class Vikhr_Nemo_instruct_pipline(IModel):
-    def __init__(self, model_name=None, device_map="auto"):
-        if model_name is None:
-            model_name = "Vikhrmodels/Vikhr-Nemo-12B-Instruct-R-21-09-24"
-
-        self.tokenizer = AutoTokenizer.from_pretrained(model_name)
-        self.model = AutoModelForCausalLM.from_pretrained(model_name, device_map=device_map)
-
-    def generate(self, prompts: list, max_new_tokens=512, return_full_text=False):
-        """Method for text generation by model.
-        Takes list of prompts and max new tokens and return list with model generated text."""
-        task = "text-generation"
-        self.pipeline = pipeline(
-            task,
-            model=self.model,
-            tokenizer=self.tokenizer,
-            max_new_tokens=max_new_tokens,
-            return_full_text=return_full_text
-        )
-        return self.pipeline(prompts)
-
-
-if __name__ == "__main__":
-    model = Vikhr_Nemo_instruct()
+        super().__init__(model_name, chat_template, system_prompt, device_map)
